chore/8845 upgrade react dnd final
