+++ src/studio/src/designer/frontend/ux-editor/containers/helpers/dnd-helpers.test.ts
@@ -1,26 +1,6 @@
-import {
-  ContainerPos,
-  getContainerPosition,
-  insertArrayElementAtPos,
-  swapArrayElements,
-} from './dnd-helpers';
-import { randomUUID } from 'crypto';
+import { getContainerPosition } from './dnd-helpers';
 import { XYCoord } from 'react-dnd';
-
-test('it should work', () => {
-  const arr: string[] = [
-    randomUUID(),
-    randomUUID(),
-    randomUUID(),
-    randomUUID(),
-    randomUUID(),
-    randomUUID(),
-  ];
-  const newnew = swapArrayElements(arr, arr[1], arr[2]);
-  expect(newnew).toHaveLength(arr.length);
-  expect(newnew[2]).toBe(arr[1]);
-  expect(newnew[1]).toBe(arr[2]);
-});
+import { ContainerPos } from './dnd-types';
 
 test('getContainerPosition returns correct positions', () => {
   const boundingBox: DOMRect = {
@@ -46,18 +26,4 @@ test('getContainerPosition returns correct positions', () => {
     const result = getContainerPosition(boundingBox, xyCord);
     expect(result).toBe(expected);
   });
-});
-
-/**
-
- */
-
-test('should insert element at pos', () => {
-  const arr = 'abc'.split('');
-  expect(insertArrayElementAtPos(arr, 'M', 9).join('')).toBe('abcM');
-  expect(insertArrayElementAtPos(arr, 'M', 0).join('')).toBe('Mabc');
-  expect(insertArrayElementAtPos(arr, 'M', 1).join('')).toBe('aMbc');
-  expect(insertArrayElementAtPos(arr, 'M', 3).join('')).toBe('abcM');
-
-  expect(() => insertArrayElementAtPos(arr, 'M', -1).join('')).toThrowError();
 });
